BACKLOG-15401: Split actions into read and write action classes
- this is necessary to make TokenPerPage work with jahia-csrf-guard
