+++ src/main/java/org/jahia/modules/jahiaauth/action/WriteMappers.java
@@ -51,12 +51,10 @@ import org.jahia.bin.Action;
 import org.jahia.bin.ActionResult;
 import org.jahia.modules.jahiaauth.impl.SettingsServiceImpl;
 import org.jahia.modules.jahiaauth.service.*;
-import org.jahia.osgi.BundleUtils;
 import org.jahia.services.content.JCRSessionWrapper;
 import org.jahia.services.render.RenderContext;
 import org.jahia.services.render.Resource;
 import org.jahia.services.render.URLResolver;
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -64,12 +62,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.*;
-import java.util.stream.Collectors;
+
 
 /**
  * @author dgaillard
  */
-public class ManageMappers extends Action {
+public class WriteMappers extends Action {
     public static final String ERROR = "error";
     private SettingsService settingsService;
 
@@ -77,20 +75,7 @@ public class ManageMappers extends Action {
     public ActionResult doExecute(HttpServletRequest req, RenderContext renderContext, Resource resource,
                                   JCRSessionWrapper session, Map<String, List<String>> parameters,
                                   URLResolver urlResolver) throws Exception {
-        JsonNode params = new ObjectMapper().readTree(req.getInputStream());
-        String action = params.get("action").asText();
-        switch (action) {
-            case "getConnectorProperties":
-                return getConnectorProperties(params);
-            case "getMapperProperties":
-                return getMapperProperties(params);
-            case "getMapperMapping":
-                return getMapperMapping(renderContext, params);
-            case "setMapperMapping":
-                return setMapperMapping(renderContext, params);
-            default:
-                return new ActionResult(HttpServletResponse.SC_OK, null, new JSONObject());
-        }
+        return setMapperMapping(renderContext, new ObjectMapper().readTree(req.getInputStream()));
     }
 
     private ActionResult setMapperMapping(RenderContext renderContext, JsonNode parameters) throws JSONException, IOException {
@@ -129,72 +114,6 @@ public class ManageMappers extends Action {
         }
 
         settingsService.storeSettings(settings);
-        return new ActionResult(HttpServletResponse.SC_OK, null, response);
-    }
-
-    private ActionResult getMapperMapping(RenderContext renderContext, JsonNode parameters) throws JSONException {
-        JSONObject response = new JSONObject();
-        if (!parameters.hasNonNull(JahiaAuthConstants.MAPPER_SERVICE_NAME)
-                || !parameters.hasNonNull(JahiaAuthConstants.CONNECTOR_SERVICE_NAME)) {
-            response.put(ERROR, "required properties are missing in the request");
-            return new ActionResult(HttpServletResponse.SC_BAD_REQUEST, null, response);
-        }
-
-        String connectorServiceName = parameters.get(JahiaAuthConstants.CONNECTOR_SERVICE_NAME).asText();
-        String mapperServiceName = parameters.get(JahiaAuthConstants.MAPPER_SERVICE_NAME).asText();
-
-        Settings settings = settingsService.getSettings(renderContext.getSite().getSiteKey());
-        Settings.Values mapperNode = settings.getValues(connectorServiceName).getSubValues(JahiaAuthConstants.MAPPERS_NODE_NAME).getSubValues(mapperServiceName);
-
-        if (mapperNode.isEmpty()) {
-            return new ActionResult(HttpServletResponse.SC_OK, null, response);
-        }
-
-        JSONArray jsonArrayMapping = new JSONArray(mapperNode.getProperty(JahiaAuthConstants.PROPERTY_MAPPING));
-        response.put(JahiaAuthConstants.PROPERTY_IS_ENABLED, mapperNode.getBooleanProperty(JahiaAuthConstants.PROPERTY_IS_ENABLED));
-        response.put(JahiaAuthConstants.PROPERTY_MAPPING, jsonArrayMapping);
-
-        if (parameters.get(JahiaAuthConstants.PROPERTIES) != null) {
-            for (Iterator<String> it = parameters.get(JahiaAuthConstants.PROPERTIES).fieldNames(); it.hasNext(); ) {
-                String property = it.next();
-                if (!mapperNode.getListProperty(property).isEmpty()) {
-                    JSONArray array = new JSONArray();
-                    mapperNode.getListProperty(property).forEach(array::put);
-                    response.put(property, array);
-                } else if (mapperNode.getProperty(property) != null) {
-                    if (property.equals(JahiaAuthConstants.PROPERTY_IS_ENABLED)) {
-                        response.put(property, Boolean.valueOf(mapperNode.getProperty(property)));
-                    } else {
-                        response.put(property, mapperNode.getProperty(property));
-                    }
-                }
-            }
-        }
-
-        return new ActionResult(HttpServletResponse.SC_OK, null, response);
-    }
-
-    private ActionResult getMapperProperties(JsonNode parameters) throws JSONException {
-        JSONObject response = new JSONObject();
-        String mapperServiceName = parameters.get(JahiaAuthConstants.MAPPER_SERVICE_NAME).asText();
-        Mapper mapperService = BundleUtils.getOsgiService(Mapper.class, "(" + JahiaAuthConstants.MAPPER_SERVICE_NAME + "=" + mapperServiceName + ")");
-        if (mapperService == null) {
-            response.put(ERROR, "Cannot find connector");
-            return new ActionResult(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, null, response);
-        }
-        response.put("mapperProperties", new JSONArray(mapperService.getProperties().stream().map(JSONObject::new).collect(Collectors.toList())));
-        return new ActionResult(HttpServletResponse.SC_OK, null, response);
-    }
-
-    private ActionResult getConnectorProperties(JsonNode parameters) throws JSONException {
-        JSONObject response = new JSONObject();
-        String connectorServiceName = parameters.get(JahiaAuthConstants.CONNECTOR_SERVICE_NAME).asText();
-        ConnectorService connectorService = BundleUtils.getOsgiService(ConnectorService.class, "(" + JahiaAuthConstants.CONNECTOR_SERVICE_NAME + "=" + connectorServiceName + ")");
-        if (connectorService == null) {
-            response.put(ERROR, "Cannot find connector");
-            return new ActionResult(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, null, response);
-        }
-        response.put("connectorProperties", new JSONArray(connectorService.getAvailableProperties().stream().map(JSONObject::new).collect(Collectors.toList())));
         return new ActionResult(HttpServletResponse.SC_OK, null, response);
     }
