+++ src/main/resources/META-INF/configurations/org.jahia.modules.jahiacsrfguard-jahia-authentication.cfg
@@ -1,0 +1,1 @@
+whitelist = *.readConnectorsSettingsAction.do, *.readMappersAction.do
