+++ src/main/resources/javascript/jahia-authentication/settings-service.js
@@ -20,13 +20,13 @@
             angular.forEach(properties, function(property) {
                 propertiesAsString += '&properties=' + property;
             });
-            return $http.get(jahiaContext.basePreview + jahiaContext.sitePath + '.manageConnectorsSettingsAction.do?connectorServiceName=' + nodeName + propertiesAsString);
+            return $http.get(jahiaContext.basePreview + jahiaContext.sitePath + '.readConnectorsSettingsAction.do?connectorServiceName=' + nodeName + propertiesAsString);
         }
 
         function setConnectorData(data, options) {
             return $http({
                 method: 'POST',
-                url: jahiaContext.basePreview + jahiaContext.sitePath + '.manageConnectorsSettingsAction.do',
+                url: jahiaContext.basePreview + jahiaContext.sitePath + '.writeConnectorsSettingsAction.do',
                 data: data,
                 headers: {'Content-Type': undefined},
                 transformRequest: function (data) {
@@ -47,7 +47,7 @@
             data.action = 'getConnectorProperties';
             return $http({
                 method: 'POST',
-                url: jahiaContext.basePreview + jahiaContext.sitePath + '.manageMappersAction.do',
+                url: jahiaContext.basePreview + jahiaContext.sitePath + '.readMappersAction.do',
                 data: data
             })
         }
@@ -56,7 +56,7 @@
             data.action = 'getMapperProperties';
             return $http({
                 method: 'POST',
-                url: jahiaContext.basePreview + jahiaContext.sitePath + '.manageMappersAction.do',
+                url: jahiaContext.basePreview + jahiaContext.sitePath + '.readMappersAction.do',
                 data: data
             })
         }
@@ -65,7 +65,7 @@
             data.action = 'getMapperMapping';
             return $http({
                 method: 'POST',
-                url: jahiaContext.basePreview + jahiaContext.sitePath + '.manageMappersAction.do',
+                url: jahiaContext.basePreview + jahiaContext.sitePath + '.readMappersAction.do',
                 data: data
             })
         }
@@ -74,7 +74,7 @@
             data.action = 'setMapperMapping';
             return $http({
                 method: 'POST',
-                url: jahiaContext.basePreview + jahiaContext.sitePath + '.manageMappersAction.do',
+                url: jahiaContext.basePreview + jahiaContext.sitePath + '.writeMappersAction.do',
                 data: data
             })
         }
